requirements.txt
streamlit<=0.78.0
Pillow<=8.1.2
pandas<=1.1.2
PyPDF2<=1.26.0
xlrd<2.0.0
docx2txt
tweepy